Adjust startPlot and volcanoPlot functions
startPlot now creates the parent directory if it does not exist.

The q-values in gene expression data are a multiple-tested p-value
giving an FDR (False Discovery Rate). The labels in the legend now
reflect that instead of falsely claiming variability of replicates
was to blame. That is a calculation taken care of in the initial
calculation of the p-value.
Also needed to fix the colors afterward for some reason.