Update logger, write.Table function and writeGOI function
In randomTools.R, added the outside and formatted parameters for when it is
 desirable to have the same string prepended and appended. Whether or not that
 string is formatted depends on what parameter it is passed in as.
    outside -> prepend and append parameters
    formatted -> formattedPrepend and formattedAppend parameters
Also in this file is logging for the write.Table function which can be turned
 off with the new verbose parameter.

In templateAnalysis.R, changed the way writeGOI function works.
It now makes files with the .txt extension instead of the (sadly) less popular
 but more appropriate .tsv extension.
Added parameter withScore, and when TRUE, will print the gene and the
 log2(fold_change) alongside it. (This is an option for GeneTrail2 input.)
Also the function now performs a check for the column name log2(fold_change).
 In the scenario it is absent, the function assumes that the local geneData
 variable is actually a geneSet type of variable (matrix of gene names to sample
 FPKMs). The file output in this case really doesn't print GOI, but just a file
 that is acceptable by GeneTrails2: See Measurements section in the GeneTrails2
documentation on Input data -
  http://genetrail2.bioinf.uni-sb.de/help?topic=input_data